backend/requirements.txt
fastapi
uvicorn[standard]
python-pptx
pillow
python-multipart
openrouter
python-dotenv
